+++ Documentation/SBV/Examples/TP/QuickSort.hs
@@ -31,7 +31,7 @@ import Data.SBV
 import Data.SBV.List hiding (partition)
 import Data.SBV.Tuple
 import Data.SBV.TP
-import qualified Data.SBV.TP.List as TP
+import qualified Documentation.SBV.Examples.TP.Lists as TP
 
 import qualified Documentation.SBV.Examples.TP.SortHelpers as SH