+++ Documentation/SBV/Examples/TP/Reverse.hs
@@ -30,7 +30,7 @@ import Data.SBV
 import Data.SBV.List hiding (partition)
 import Data.SBV.TP
 
-import qualified Data.SBV.TP.List as TP
+import qualified Documentation.SBV.Examples.TP.Lists as TP
 
 #ifdef DOCTEST
 -- $setup